Fix ShellCheck errors

+++ setup-scripts/scripts.sh
@@ -2,6 +2,7 @@
 #
 # Installs scripts
 
+# shellcheck source=../config/scripts/lib/utils.sh
 source "$HOME/.dotfiles/config/scripts/lib/utils.sh"
 
 echo_info "Installing scripts..."
@@ -10,7 +11,7 @@ dotfile_dir="$HOME/.dotfiles/config/scripts"
 config_dir="$HOME/.local/scripts"
 
 # Cleanup to avoid broken links hanging around
-rm -r "$config_dir/*" 2>/dev/null
+rm -r "${config_dir:?}/*" 2>/dev/null
 
 for dotfile in "$dotfile_dir"/*.sh; do
   link_file="$config_dir${dotfile#"$dotfile_dir"}"
